Intial Agent remote implementation
So, here is the start for Guest Agent support. Savanna communicates with
the agent via oslo.messaging. The agent's code will be committed later in
a separate repository.

One remark: current code assumes that the agent is not pre-installed on
the image, so additional userdata is added. The userdata installs
and starts the agent.

Current limitations:
  * Works for Vanilla and HDP plugins
  * Works on Ubuntu and CentOS guests
  * Works with Rabbit MQ only

Note: the guest agent repo is at
https://github.com/stackforge/savanna-guestagent

Correspondingly the project name in gerrit is stackforge/savanna-guestagent

Implements: blueprint guest-agent
Change-Id: I787b6b058989568553815bc6569e6fe8f46df6b5