Fix resource discovery for datasources
The lookahead code which checks path to source files and ssl certificate
did not properly consider the shared resources when package is installed.
Now data source files are looked locally first and then, if not found and
the path is relative, inside the default templates directory.
The default paths for test resources have been then changed to use defaults/
as base directory, with the exception of unit tests.
As the file referenced by yaml files must now exist, a new dummy file .crt
file was added for unit tests.

Change-Id: Ic385c6b4d694d95a11a626a3107fa1906924bd67