Run the tests inside the test directory
This change allows the test runner to be executed everywhere without
having a .testr.conf file. A copy of .testr.conf file (which is now
properly released and installed) is copied to the test directory, which
is now used as working directory for the ostestr call.
The path to the directory where the execution has been started is now
passed to the test template, so that the code can write there data like
the test reports, the key (if the resources are kept), etc.

Change-Id: I16ddba2ad6e26381ffc4b447666f38f2a7740ecb