New Pig example with a User Defined Function
Replace the existing Pig/UDF example with an original one.
The license of the replaced example was not totally clear.

The input and expected output of the old example has been kept as they
are used also by other tests.

Even if we are working towards the removal of jar file from this repository,
rebuilding Hadoop-based jars is not trivial, so the jar file (compiled with
target JVM 1.6) is part of the patch.

Change-Id: Ib86f63458797dc10b19334177dab01e16894ca57