Migrate the imported client tests to the Tempest plugin
The imported tests will be executed with the other Tempest tests.
Those are special Tempest tests: they use Tempest-as-a-library (including
some classes with unstable API which should be migrated to the solutions
now available in tempest.lib), but they use the official clients.

Make sure that the required parameters are set in the Tempest configuration
file (or reduce their need when possible).

Change-Id: I6b5cf378f7a6c3ae83c25874f1c9185cc0a03e81