Get plugin data: dynamically discover the output file
The format of the plugin file changed post-Newton (it includes also the
version number now), so instead of trying the various possible combination,
read the name returned by the command.
Fallback name: the new behavior (in case the output will be removed, which
can't happen for the already released versions).

Also, cleanup the file with the plugin data after the test.
Change-Id: I3d4a4005eda74b2c1c35b1f23b9ce27a1797c9c4