Test templates: add the ocata/ dir, remove liberty/ etc
The current set matches the job tested on the Sahara CI on the
current master, which is going to become Ocata.
Remove the directory with Liberty templates (Liberty has been EOL
for a while).
Also remove mapr-5.0.0.mrv2.yaml.mako in the main test directory
(it is going to be removed completely in newer versions).
Finally, fix the unit tests which depended on the test templates
for Liberty (use Newton for now).

Change-Id: I61cae34c7ce3b7de22502f4ebb66464df9164647