Fix sahara tempest plugin configuration
config.CONF cannot be used in register_opts, because
register_opts is invoked by Tempest as part of the process
of building config.CONF itself, so that ends up in a
recursive call that raises duplication opion exceptions.

Make sure that register_opts uses the supplied conf
object (as opposed to tempest.config).
Register the "sahara" option in the existing
'service_available' group rather than recreating the
group.

Cleanup all the exception handling that was required
to catch the issues described above. Any more exception
should surface as it would be sympthom of an issue to
be fixed.

Depends-on: I1afae0d1f9c4a6aec2742aaba4928cdde54b224f

Change-Id: I1a2cf1cfe367eaf86eb6844ab6bfd8d3c2de51ec