doc: switch to openstackdocstheme, add metadata, fixes
Move doc/ and releasenotes/ to openstackdocstheme
and remove the dependency on oslosphinx.
Initialize the correct link to the bug tracker (logABug).

Also, fix the "Indices and tables" section: don't set it as top-level
section (only one seems to be supported by the new theme) and remove
a dead link (which probably never worked).

Change-Id: I547f4f4578b39c5c1a0b3bf134913f4c82284162