Remove a useless call in tempest tests and add .stest.conf
.stestr.conf is needed to properly discover the location of the
unit tests, otherwise also the tempest tests are executed (there
is a known issue in ostestr when .stestr.conf is not used).
.testr.conf is kept for now for the coverage tests.

Notes about the cls.set_network_resources() call inside
resource_setup():
- its usage in this place triggers an error, but without behavioral
  changes: if it was not working, it was not working even before;
- a call without parameters actually resets the network settings,
  which is maybe not the intended behavior.
So it should be just safe to remove this call, and if the behavior
does not change, this Tempest plugin with the fix should work also
with older versions of Tempest.

Change-Id: If28ecafde23ba4ed54fef5c4a4588dc77370da9c