Fix: plugin installed does not mean service available
The patch that introduced this configuration option set the default
availability of the service to True, but this is not correct (even if
I approved that patch).
The plugin can be installed but it does not mean that the service is
automatically available: for example if all plugins are installed
in a pre-built tempest image (container, VMs, etc), or just for
testing.
Users are expected to tune tempest.conf anyway. It is worth noting
that the devstack plugin for Sahara sets the value of the key to True,
while TripleO uses python-tempestconf which discovers the value.

Change-Id: I878d99ab7a65d1082eed6c1cdc650099bcf75ddc