Fix lintstack.py with Python3
The pylint target was switched to python3, but I didn't realize
that it's not really working. This change should make lintstack.py
compatible with Python 3.

Related note: we should find a unified solution for all repositories
which have a local copy of lintstack.py/lintstack.sh instead of
duplicating this code over and over.

Change-Id: I5727e33f8ac40347e47e4f5f4e64579897a66f6f