Allow to change the version of API also for CLI tests
The tempest.conf configuration key data-processing.saharaclient.version
was renamed to api_version_saharaclient; the old name is still accepted.
More important, the variable controls also the version of the API
used for CLI tests, not only for tempest.lib-based client tests.

The signature of ClientTestBase.openstack changes a bit (no more *args
and **kwargs) but it should not cause any regression, because the
method only passes the parameters to the corresponding function
from tempest.lib.cli, which accepts that specific set of arguments.

Change-Id: I2034591f9f336100b2eb3c0e42e4b94227c90625