APIv2 support for all Tempest-based tests
The same tests can be used against APIv1.1 or APIv2, but
not at the same time.
This Tempest plugin provides now a DataProcessing Tempest client
for APIv2.

Most of the tests could be ported and used with minimal changes
with APIv2, with the exception of api.test_jobs; the corresponding
new tests are under api.test_job_templates.

The tests are executed against APIv2 when a certain configuration
key is set into tempest.conf:
- data_processing.use_api_v2 for API tests (based on the Tempest client);
- data_processing.api_version_saharaclient for client and CLI tests
  (both based on python-saharaclient).

A new job (sahara-tests-tempest-v2) runs the APIv2 tempest tests.

Depends-On: https://review.openstack.org/584110
Depends-On: https://review.openstack.org/581774
Story: 2004511
Task: 28237
Task: 28238
Change-Id: Ic33d4c5c550d093b305ee5931feb9f697df04ba7