New job for sahara-scenario based on Python 3
Only sahara-scenario is executed using Python 3;
the deployment itself (devstack) still uses Python 2.

In order to ease the transition, add a venv-py2 tox target
that can be used by the sahara and python-saharaclient jobs
on the older branches to still use python2 when the main
venv will be switched to python3.
This venv is used to make sure that the old branches
are still using Python 2 regardless of the configuration
of sahara-tests-scenario.

After this change is merged, the scenario jobs on older branches
should be changed to explicitly use venv-py2. When this is
done, sahara-tests-scenario-runner-py3 can be removed and
sahara-tests-scenario can be switched to use venv-py3.

Change-Id: I379a993e45da842cea24e50afe84c89556c0e700