APIv2 jobs also on the gate queue, add the py37 job
Enable the API v2 tests also on the gate queue, in addition
to the check queue.

Add the py37 unit test job and complete the list of the
available virtual environments in the envlist key of tox.ini.

Remove the usage of os_image and os_nova_flavor: Zuul now uses
Ansible 2.7, which requires a version of openstacksdk not available
in queens and pike.

The addition of the py37 job implements the following task:
Story: 2004073
Task: 27450
Change-Id: I016f674b3bc3e4617ee724261e6a0c22d8520fbc