Update api-ref location and some links
The api documentation is now published on docs.openstack.org instead
of developer.openstack.org. Update all links that are changed to the
new location.

Note that redirects will be set up as well but let's point now to the
new location.

For details, see:
http://lists.openstack.org/pipermail/openstack-discuss/2019-July/007828.html

At the same time, fix the links that have been broken by
the separation between API v1.1 and v2.

Change-Id: Ic72c0cd85d3d42a62071825c242df599d2442f99