Py2 cleanup: py3 on old branches (or none), remove py27
Make sure that the job running on master are py3 based,
according the established policy.
RGW must be used to provide the Swift API on Stein,
as devstack cannot properly deploy Swift/py2
in a Python 3 environment (old issue never fully solved).

Remove all EM releases where Python 3.6 is not available
and Tempest is not supported anymore. They will be gated
by sahara and python-saharaclient stable gates if needed;
the py2 variants are kept for this reason.
At the same time, remove the py35 unit test job,
as it is not supported anymore.

If an incompatible changes breaks an older branch,
an older sahara-tests version will be pinned there.

Remove the leftover py27 unit test jobs, which were
still needed when the py2 version of sahara-tests was
used on older branches (the py3 version is always used
even when the deployment is py2).

Finally, fix a multi-line issue with a shell command
which strangely happens for the master job only.

Change-Id: Id80ef7e186f6328ab162b92eebababd7e4886560